Bug fixes; incorrect indexing
Fixed several indexing issues.

In some cases (likely depending on Zotero version),
body is a either an Array of objects or an Array
of Array of objects, so convertTo methods were
fixed to deal with either case.

Additionally, convertTo methods incorrectly
produced bodies that were Array of Arrays
instead of Arrays.

Change-Id: I212abccef0f37584c6e3c2ee84293c55476ec237