Bug fixes: fix status response codes
Response codes for various errors were inconsistent
or non-existant in areas.

* Fixed bug T8884 where response codes from scraped
URLs were ignored (Line 46 in scrape.js). Now returns basic
citations for these URLS (i.e. with URL in title) instead of
scraped titles and a 520 error to indicate the resource wasn't found
or had errors.
* Fixed bug T78388 where failure to export to bibtex format
returned a 200 OK and an empty body. Now returns a 404
not found error and a message saying the requested format
couldn't be obtained.
* Changed scrape() in scrape.js to take a custom responseCode
as second argument in callback for more nuanced responseCodes.
* Changed zoteroExportRequest() in zotero.js
to take responseCode as second argument in callback instead of
full response.

Bug: T88884
Bug: T78388
Change-Id: I9b94fe40bfe71fadc5d227da50eb142d285b14b2