Fix acceptLanguage commit a30d6be
* Fixed errors to CitoidService
requestFromDOI method.

* Added test for requestFromDOI
'doi'

* Fixed test from acceptLanguage commit
'German Twitter' which checked for
incorrect responseCode

* Throw callback errors in all tests

* Check error in test 520 (now test
ENOTFOUND) and throw if the error is not
the expected error (getaddrinfo ENOTFOUND)

* Add test 404 for cases where a 520 error
is expected but no error should be thrown
(for when there is a resource at the server
but the HTTP response is 404).

* Increase timeout default for mocha
since the 'doi' test takes longer
than the previous timeout length.

Change-Id: I93f8c699e7f58375e4fa8dd7c48dd721969e3659