Add field "websiteTitle" in itemType "webpage"
Field "publicationTitle" was erroneously being
used instead of field "websiteTitle" in the native
scraper for items of type "webpage". This is a
temporary change in a backwards compatible way
until the erroneous field can be removed safely.

* Fix translators/openGraph.js to have the correct
field "websiteTitle" instead of "publicationTitle".

* Convert backup behaviour for publicationTitle
in lib/Scraper.js to create field websiteTitle
instead of publicationTitle

* Add temporary statement to copy value of field
websiteTitle to field publicationTitle for
backwards compatibility until TemplateData
can be updated on all wikis.

* Add test ensuring both fields websiteTitle
and publicationTitle are present in natively
scraped "webpage" citations.

Bug: T93513
Change-Id: I2c9d81cea144ddd3bc5bac917cbbeb629d5aa608