Change how pubmed and pmcs are requested
The PubMed ID converted API was frequently
not returning results for valid PMIDs, so
instead we request metadata from pubmed
URLs directly.

CitoidService

* Remove requestFromPubMed, which used
the ID converter API to get a DOI
* Replace with requestFromPMCID and PMID
which concatenates id onto a PubMed URL,
verifies the server sends a 200 ok
response code, and sends to
requestFromPubMedURL.
* Add requestFromPubMedURL which sends
pubmed urls to Zotero first, and to
native scraper second. If the scraper
detects an HTTP error, it returns 404,
otherwise it scrapes the page.

Tests

* Convert pmid test for to a PMID which has
no results in the converter API but has a
valid URL.
* Add test for PMC with PMC prefix
* Add test for PMC without PMC prefix
* Add test for invalid PMCID
* Add test for invald PMID

Bug: T93335
Change-Id: Ie2830d42fa63fecd30db72f723e6d5d9979f51b5