Check for empty arrays returned by Zotero
Zotero is a funny beast. Sometimes there is nothing to be scraped, or
its translators cannot translate properly, so it returns an empty array
or nothing at all. The joke is on us, though, because knowing its nature
we have trusted it. We don't any more. This patch hot-fixes that. A
proper solution is in the works, this commit only helps Citoid not to
crash in production.

Change-Id: I0b91be17a2aa2950914510724864ecd245cc5059