Use the same cookie jar throughout a request's life
unshorten() needs to have a cookie jar as well since some sites, such as
ScienceDirect, do not allow the redirects to continue before seeing the
cookie they sent. Furthermore, move the cookie jar creation to
CitoidRequest, so that the scraper can later reuse these cookies and
thus scrape the resource faster.

This partially addresses T97042 in that it avoids redirect hell caused
when unshorten() is trying to find the final URL to scrape.

Bug: T97042 (partially only)
Bug: T106975
Change-Id: Id857d7eebb0d077408d6a4a4c86e3d2ab864ed5f