New dublinCore translator and template translator
* Create template translator for new translators
(template.js.txt)
* Move utility functions into util folder (were
previously in general translator which refers
to a specfic metadata type, not just general
functions.)
* Improve existing translators to cover all
possible citation itemTypes values.
* Create a translator for dublinCore.
* Add logic to extract DOI from DC metadata
and use to get more metadata from crossRef.
* No longer translate date from crossRef as
it only ever contains the year and this results
in it being converted to January 1st in ISO
which is typically wrong.

Bug: T108175
Bug: T76224
Bug: T113232
Change-Id: Ieb380b7ae1da98ccbd0a9a735c5dcbf48c5270c8