Fix lack of response when export fails
During failed export, no error was being
returned because fillBody function was
only filling the body with errors that
occured before export. This fills the body
with errors that occur after export.

Unfortunately, for some reason this
response is being returned as a Buffer
and therefore acts up in the browser,
although it demonstrates the correct
behaviour in curl.

Bug: T115271
Change-Id: I09e40682521f4d99775c4669238117b813985521