Add bepress/highwire tag support
* Fix bug in unit tests for scraper that
caused all the tests to pass if the value of
the translator was undefined
* Fix uncaught type errors in openGraph
translator
* Change make translator utilities to take
entire metadata object instead of only
one property value so multiple property values
can be used (i.e. start page and end page
to create pages value.)
* makePagesTranslator function added to translator
utilities + unit tests
* Fix whitelist of fields in Zotero fcns to
include itemType and exclude whitelisting of
creators for note itemType only
* Make ISSN/ISBN in coins translator generalisable
and move to utils.js
* Move test from scraping/index.js to noZotero
as the most recent translators support that journal
and so the results are coming from Zotero instead
of the native scraper.
* Use highwire and bepress doi info to find doi
* Use bepress/highwire attributes to assign type

Bug: T123361
Change-Id: Id3db251b66141a6fd6c447bde536ef62649be84a