Add source field in mediawiki format
Add 'source' field with value type Array to mediawiki format.

This indicates the source of the retrieved metadata. This
is returned in an Array as multiple sources of metadata
may be present. This currently includes 'WorldCat',
'Crossref', 'Zotero', and 'PubMed' where the PMC/PMID to DOI
API is used, and 'citoid' when citoid uses metadata it natively
scrapes from the page. If 'source' is not present in the response
this is usually a 520 response, where the page is unable to be
scraped, and queries to third parties have failed as well.
Zotero results may include metadata retrieved from CrossRef,
WorldCat, and others.

* Move multiple tests which previously were not in Zotero
to test/features/scraping/noZotero.js as they are now
in Zotero and need Zotero to be disabled to work
* Check source type for multiple scraper tests.
* Add assert function to check if a string is present in an
Array.

Bug: T145465
Change-Id: I44eec8a77f3c0893aefd917261c4423b49fe17bc