Remove deprecated endpoint and format
Remove url endpoint which pre-dates tests and we
do not test it. It has been deprecated
for ~3 years. Restbase also does not
support this endpoint.

Remove mwDeprecated format which is of
similar age, and fix tests to now interpret
the format as invalid.

Update packages &
Bump major version as these changes are not
backwards compatible.

Unable to update body-parser as it causes
ENETUNREACH errors when resolving DOIs for
some reason.

Bug: T93514
Change-Id: Idd3ca4f03480ca8bf3ffbdb92d9459be5b2977ac