Fallback to crossRef for restricted URLs
Previously requests to restricted URLs would
immediately fail; in this change, if there
is a DOI present, we allow the request to
succeed if we are able to get the data
from the DOI from crossRef.

This is done by allowing the request to
proceed to lib/Scraper.js and just checking
that the hostIsAllowed there as well.

Bug: T165105
Change-Id: I811f48ddc021b8e4c32f82e5d9f16d079fe5003a