Try to improve performance with pubmed
Adds config option to not request pubmed
after every request.

If pubmed: true, then lib/Exporter will wait for the initial
request for additional identifiers (if made) and attempt
to make a new request for identifiers later if not made.

If pubmed: false, then the identifiers are requested
at the beginning of the request, when the DOI
is known either because the DOI, PMID or PMCID was
initially requested or was found from the URL.
We do not wait for the pubmed request to
complete, so if the request takes longer than the time it
takes to finish filling in the citation from other sources,
then we simply return without it.

Bug: T162886
Change-Id: I6b5a381ed46cef6958cb2dcd37504f2bbc41eff1