Bump html-metadata, fix unshorten bug, fix tests
* Increase html-metadata version to 1.7.0
* Fixed title with upstream change in test/features/scraping/lang.js
* Formatting in test/features/scraping/zotero.js
* Pubmed conf test wasn't functioning because extend doesn't respect
undefined; switched to Object.assign instead which does.
* Skip chronicalingamerica url; dead upstream.

In some cases, we were following redirects without checking if the
host was allowed. This was causing some tests to fail,
but only recently for some reason. This change replaces
followAllRedirects in lib/Scraper with unshorten. The behaviour
of lib/unshorten.js was changed to succeed with non-shortened URLs
for checking whether hostIsAllowed for all urls in a chain.

Bug:T185694
Change-Id: I9e63b450e7cd4853537365aeefe5a748e1d20c93