Add open search capability with crossref
This adds the ability to use arbitrary search
terms (such as the title of a work, or a fully
formatted citation) using crossref.

This works well for journal articles
but not for books. The plan is to add worldcat
search after this (T162357) so it will work
for books as well.

As a consequence, some search terms that were once
disallowed (such as queries that contain <>) are
now allowed to be passed on to crossref. Thus,
these tests return 200 instead of 400.

In this change, we only return the first result
from crossref, although multiple results are
available. As many as two results may be returned
in total if the query also contains a url.

Bug: T179123
Change-Id: I29218cac0fd29cde44858ae4869f446503a8a4ba