Force early return for some DOIs
Set timeout in config for DOIs.

If the timeout has occured and we have crossref data but
no data from URL, this will return only crossref data.

If we get back both crossRef and
fromURL in under wait time, it will have data from both.

If we don't have either
crossref or data from the URL by the time the timeout promise
is complete, it will continue to wait until one of them resolves and
return with whichever is first.

Bug:T165105
Change-Id: I46346c0087fdf5a3f95ce543a88c904bc7fe8139