Update citoid to work with new translation-server
Update ZoteroService.js to work with new
translation-server, located at
https://gerrit.wikimedia.org/r/#/admin/projects/mediawiki/services/zotero
which is a mirror of
https://github.com/zotero/translation-server

This change is not backwards compatible with the old
version of translation-server, located at
https://github.com/zotero/translation-server-old and
https://gerrit.wikimedia.org/r/#/admin/projects/mediawiki/services/zotero/translation-server

Update tests. Contains some regressions. Notably, the
urls that previously went to citoid now go to zotero,
causing some issues we fixed locally to no longer be
fixed, such as a few i18n issues. Going forward we
should fix those issues in zotero instead.

The new zotero also has their own "search" feature, but
for the time being we have not enabled this, only the
URL capabilities.

Bug:T197242
Change-Id: Ib3cff5e7eda2a7e61efb9944c86f7cf9d9604ecb