Remove xISBN and replace with zotero
The xISBN service was decomissioned by
worldcat in 2016 and finally went down
for good in 2018. Removing legacy
code relating to this service.

Replace with zotero search for isbn.
This allows third party hosts to search
for isbn.

Remove worldcat key from dev as it
expired in November and will not be
renewed. Skip tests that require this
key. To test these features you will
need to request your own sandbox key.

This change does not affect results
in production as it only uses zotero
if wskey is not enabled. In the
future we may wish to merge these
results instead.

Bug: T214802
Change-Id: I0fc4b7a169c4c37327437da924f02f56a03ee3ac