Clean up final traces of xisbn service
xisbn was a WorldCat service that went offline
several years ago. We previously removed most
code relating to it in:

a5459ac0bc89

This removes the option to use it from all the
configs, one line loading the config that was
missed in a prior clean-up,
and an incorrect comment that referred to it.

Also removes two xisbn specific testing files.

Also removes a wskey key that shouldn't be
there.

Change-Id: I2b605178ba706421166919fac8f3d0439af7de0d