Improve error handling
Fix up confused logic between the responseCode of the total
response, and responseCode used for each Citation.

Previously, Citations would have responseCodes, but this
is nonsensical because only a CitoidResponse should have
a response code.

Now the responseCode is set for each CitoidResponse only
once (in fillBody), and each Citation has a CitoidError relating to
the error message or error code for that particular Citation.

Validity of each Citation is checked by checking if it has
a CitoidError.

Change-Id: Ib773fbec705d2ce6dfb9f9cd2babda2834d3326e