Avoid logging full err for warns
HTTPError 'msg' field contains full body and was causing
issues for logstack.

Fixes outgoingReqResult serialiser
to not include err.msg (which can potentially be
large as well) and avoid adding duplicate fields.

This is no longer causing issues due to switching
to get requests so the body is longer in the message
but will future proof it if the serialiser is used
for get requests again.

This also logs fewer duplicate fields, so fewer are
required to be indexed in logstash.

Follow up to I83d80e7db
Bug: T364901
Change-Id: I6aba9f177b6c3b9e45346afada3db87c0de41a31