Add 30s bucket to citoid_router_request_duration_seconds
Why:
While running the SLO exercise for Citoid, we realized that we don't
have good historical data for our percentiles because our top buckets is
1 second. If we want to have a metric for the experience of end users
regarding latency, we need more buckets

What:
Add a 30 second bucket at the top of the list. That bucket is
substantially larger than the previous one, which is 1 second and is
sufficient for the p70 (most of the time), but insufficient for anything
above that. 30 is chosen as a good round number and taking into account
we have some 20+ second timeouts. We won't be adding more buckets
between the penultimate and last bucket, to preserve disk space, unless
a need arises, which we 'll know later in time

Change-Id: I6faacd23dde8ced0b29464ca2bb930358f44c8d0