Add headers to access-control-allow-headers
Add user-agent, accept-language, and referer to
the access-control-allow-headers as

a) accept-language is used in the app itself,
and is listed as one of the allowed headers in restbase.

b) We log user-agent in turnilo, so this is useful if supplied
by the consumer. 

c) Referer, as this is logged by turnilo.

Bug: T361576
Change-Id: Id3aa5ae0ffb95b487b53803a2999992116d115f9