Redo "Remove preq from ZoteroService"
Redo of I4ce142e4

* Error handling differed from preq and this *may*
have have resulted in the increase in 500 internal server errors.
Attempt to resolve this by throwing HTTP errors and wrapping
in a bluebird promise.

* Add tests to ensure errors are handled correctly
when exporting into a different format. Unfortunately
these tests don't manage to replicate the production 500s, but
we're now logging details of 500s (done in  I4a576fa4528c7)
so if this *doesn't* fix the issue,
it'll be easier to debug. 

Bug: T361602
Change-Id: Id293804f1e2d320b17c1b0ca57a67520e750bcec