Fix broken date validation
Chrono-node date parsers would fail when the current day
didn't exist in the month in the date being parsed for partial dates.
This is because it would set the "implied" day to the current day
and then the date would be invalid if it were i.e. February 30th.

This sets the implied day in the parser to the 1st since this is
guaranteed to exist in any given month.

Change-Id: I90c4c35bcc84d36b92e4e5b39d0f6f7e76776186