Fix empty citation bug
Some input was causing empty citations with
access-date only to return, when the input
contained no spaces.

This was causing an error when in some cases
the input matched this criteria but no url
was extracted; now checks the url exists
before sending to requestForURL.

Add test for this edge case.

Change-Id: Icce2a64f4dc5a26c079284cea6b532466ea3b095