Replace string.match() with RegExp.test() where possible
The main motivation here is to make the code state more clearly
what it's actually meant to do. The .match() method returns an
array of matches. But this code is not interested in the actual
match, only if a match happened (true/false). That's what .test()
is for.

Tests prove that no behavior changed.

I'm also removing the 2000 character length limit in one case
because why should an archive link that's longer *not* be detected
as an archive link any more?

Change-Id: Ia1e418d77521fc11dc30619b1f6531f380be3cff